Build TestCase for login and signup routes
[Delivers #161243594]
